+++ api/tests/certification/session-management/unit/application/certification-candidate-controller_test.js
@@ -60,4 +60,24 @@ describe('Unit | Controller | certification-candidate-controller', function () {
       expect(response.statusCode).to.equal(204);
     });
   });
+
+  describe('#endAssessmentBySupervisor', function () {
+    const certificationCandidateId = 2;
+
+    it('should call the endAssessmentBySupervisor use case', async function () {
+      // given
+      sinon.stub(usecases, 'endAssessmentBySupervisor');
+      usecases.endAssessmentBySupervisor.resolves();
+
+      // when
+      await certificationCandidateController.endAssessmentBySupervisor({
+        params: { id: certificationCandidateId },
+      });
+
+      // then
+      expect(usecases.endAssessmentBySupervisor).to.have.been.calledWithExactly({
+        certificationCandidateId,
+      });
+    });
+  });
 });
